move test utils to separate file
